install stream-combiner module
The stream-combiner module creates a pipeline from a list of streams,
returning a single stream that exposes the first stream as the writable
side and the last stream as the readable side like the duplexer module,
but with an arbitrary number of streams in between. Unlike the duplexer
module, each stream is piped to the next.